\name{sample.traits}
\alias{sample.traits}
\docType{data}
\title{Dummy Trait Data}
\description{
Dummy trait data generated using the R/qtl package.
}
\usage{data(sample.traits)}
\format{
  The format is:
 num [1, 1:20] 0.574 1.159 0.253 -0.585 -0.611 ...
}
\source{Broman,K., et al. (2003) R/qtl: QTL mapping in experimental crosses, Bioinformatics, 19, 889-890.}
\examples{
data(sample.traits)
}
\keyword{datasets}
