\name{WCQ-package}
\alias{WCQ-package}
\alias{WCQ}
\docType{package}
\title{WCQ: QTL Mapping with Small Sample Size}
\description{Contains implementation of WCQ QTL mapping method as well as the Chen-Qin two-sample and one-sample test of means}
\details{
\tabular{ll}{
Package: \tab WCQ\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2012-07-29\cr
License: \tab GPL-3\cr
}
}
\author{
Jan Michael Yap

Maintainer: Jan Michael Yap <jcyap@dcs.upd.edu.ph>
}
\references{Chen,Y.X., Qin,Y.L. (2010) A Two-Sample Test for High-Dimensional Data with Applications to Gene-Set Testing, The Annals of Statistics, 38(2), 808-835.}
\keyword{ QTL mapping }
